Update home
Chia các component ra hợp lý không còn rối như trước nữa
