+++ Shopping-React/frontEnd/src/components/Aside/Aside.tsx
@@ -1,162 +1,37 @@
 import React from "react";
+import Input from "../Input";
+
+import { Star1, Star2, Star3, Star4, Star5 } from "../Star/Star";
+import Allfile from "./components/Allfile";
+import Category from "./components/Category";
+import Search from "./components/Search";
+import Price from "./components/Price";
 
 export default function Aside() {
   return (
     <div className="py-4">
-      <a className="flex items-center font-bold">
-        <svg viewBox="0 0 12 10" className="mr-3 h-4 w-3 fill-current">
-          <g fillRule="evenodd" stroke="none" strokeWidth="1">
-            <g transform="translate(-373 -208)">
-              <g transform="translate(155 191)">
-                <g transform="translate(218 17)">
-                  <path d="m0 2h2v-2h-2zm4 0h7.1519633v-2h-7.1519633z"></path>
-                  <path d="m0 6h2v-2h-2zm4 0h7.1519633v-2h-7.1519633z"></path>
-                  <path d="m0 10h2v-2h-2zm4 0h7.1519633v-2h-7.1519633z"></path>
-                </g>
-              </g>
-            </g>
-          </g>
-        </svg>
-        "Tất cả danh mục"
-      </a>
+      <Allfile />
+      <Category />
       <div className="my-4 h-0.5 bg-gray-300"></div>
-      <ul>
-        <li className="py-2 pl-2 hover:text-gray-500">
-          <a className="relative px-2">Áo thun</a>
-        </li>
-        <li className="py-2 pl-2 hover:text-gray-500">
-          <a className="relative px-2">Đồng hồ</a>
-        </li>
-        <li className="py-2 pl-2 hover:text-gray-500">
-          <a className="relative px-2">Điện thoại</a>
-        </li>
-      </ul>
-      <a className="mt-4 flex items-center font-bold uppercase">
-        <svg
-          enable-background="new 0 0 15 15"
-          viewBox="0 0 15 15"
-          x="0"
-          y="0"
-          className="mr-3 h-4 w-3 fill-current stroke-current"
-        >
-          <g>
-            <polyline
-              fill="none"
-              points="5.5 13.2 5.5 5.8 1.5 1.2 13.5 1.2 9.5 5.8 9.5 10.2"
-              stroke-linecap="round"
-              stroke-linejoin="round"
-              stroke-miterlimit="10"
-            ></polyline>
-          </g>
-        </svg>
-        Bộ lọc tìm kiếm
-      </a>
+
+      <Search />
       <div className="my-4 h-[1px] bg-gray-300"></div>
       <div className="my-5">
-        <div>Khoan gia</div>
-        <form className="mt-2">
-          <div className="mb-2 flex items-start">
-            <div>
-              <div className="false py-1 default-input">
-                <input
-                  className="w-full appearance-none bg-transparent leading-tight focus:outline-none grow"
-                  name="price_min"
-                  type="text"
-                  placeholder="₫ TỪ"
-                  value=""
-                />
-              </div>
-            </div>
-            <div className="mx-2 mt-2 shrink-0">-</div>
-            <div>
-              <div className="false py-1 default-input">
-                <input
-                  className="w-full appearance-none bg-transparent leading-tight focus:outline-none grow"
-                  name="price_max"
-                  type="text"
-                  placeholder="₫ ĐẾN"
-                  value=""
-                />
-              </div>
-            </div>
-          </div>
-          <button
-            type="submit"
-            className="flex items-center outline-none transition duration-300 bg-primary text-white flex items-center space-x-2 rounded-md px-4 py-2 hover:bg-opacity-80"
-          >
-            Áp dụng
-          </button>
-        </form>
+        <Price />
         <div className="my-4 h-[1px] bg-gray-300"></div>
         <div className="text-sm">Đánh giá</div>
-        <div className="my-3">
-          <div
-            className="flex items-center text-sm"
-            aria-hidden="true"
-            role="button"
-          >
-            <svg
-              stroke="currentColor"
-              fill="currentColor"
-              stroke-width="0"
-              viewBox="0 0 1024 1024"
-              className="h-5 w-5 fill-yellow-400"
-              height="1em"
-              width="1em"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M908.1 353.1l-253.9-36.9L540.7 86.1c-3.1-6.3-8.2-11.4-14.5-14.5-15.8-7.8-35-1.3-42.9 14.5L369.8 316.2l-253.9 36.9c-7 1-13.4 4.3-18.3 9.3a32.05 32.05 0 0 0 .6 45.3l183.7 179.1-43.4 252.9a31.95 31.95 0 0 0 46.4 33.7L512 754l227.1 119.4c6.2 3.3 13.4 4.4 20.3 3.2 17.4-3 29.1-19.5 26.1-36.9l-43.4-252.9 183.7-179.1c5-4.9 8.3-11.3 9.3-18.3 2.7-17.5-9.5-33.7-27-36.3z"></path>
-            </svg>
-            <svg
-              stroke="currentColor"
-              fill="currentColor"
-              stroke-width="0"
-              viewBox="0 0 1024 1024"
-              className="h-5 w-5 fill-yellow-400"
-              height="1em"
-              width="1em"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M908.1 353.1l-253.9-36.9L540.7 86.1c-3.1-6.3-8.2-11.4-14.5-14.5-15.8-7.8-35-1.3-42.9 14.5L369.8 316.2l-253.9 36.9c-7 1-13.4 4.3-18.3 9.3a32.05 32.05 0 0 0 .6 45.3l183.7 179.1-43.4 252.9a31.95 31.95 0 0 0 46.4 33.7L512 754l227.1 119.4c6.2 3.3 13.4 4.4 20.3 3.2 17.4-3 29.1-19.5 26.1-36.9l-43.4-252.9 183.7-179.1c5-4.9 8.3-11.3 9.3-18.3 2.7-17.5-9.5-33.7-27-36.3z"></path>
-            </svg>
-            <svg
-              stroke="currentColor"
-              fill="currentColor"
-              stroke-width="0"
-              viewBox="0 0 1024 1024"
-              className="h-5 w-5 fill-yellow-400"
-              height="1em"
-              width="1em"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M908.1 353.1l-253.9-36.9L540.7 86.1c-3.1-6.3-8.2-11.4-14.5-14.5-15.8-7.8-35-1.3-42.9 14.5L369.8 316.2l-253.9 36.9c-7 1-13.4 4.3-18.3 9.3a32.05 32.05 0 0 0 .6 45.3l183.7 179.1-43.4 252.9a31.95 31.95 0 0 0 46.4 33.7L512 754l227.1 119.4c6.2 3.3 13.4 4.4 20.3 3.2 17.4-3 29.1-19.5 26.1-36.9l-43.4-252.9 183.7-179.1c5-4.9 8.3-11.3 9.3-18.3 2.7-17.5-9.5-33.7-27-36.3z"></path>
-            </svg>
-            <svg
-              stroke="currentColor"
-              fill="currentColor"
-              stroke-width="0"
-              viewBox="0 0 1024 1024"
-              className="h-5 w-5 fill-yellow-400"
-              height="1em"
-              width="1em"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M908.1 353.1l-253.9-36.9L540.7 86.1c-3.1-6.3-8.2-11.4-14.5-14.5-15.8-7.8-35-1.3-42.9 14.5L369.8 316.2l-253.9 36.9c-7 1-13.4 4.3-18.3 9.3a32.05 32.05 0 0 0 .6 45.3l183.7 179.1-43.4 252.9a31.95 31.95 0 0 0 46.4 33.7L512 754l227.1 119.4c6.2 3.3 13.4 4.4 20.3 3.2 17.4-3 29.1-19.5 26.1-36.9l-43.4-252.9 183.7-179.1c5-4.9 8.3-11.3 9.3-18.3 2.7-17.5-9.5-33.7-27-36.3z"></path>
-            </svg>
-            <svg
-              stroke="currentColor"
-              fill="currentColor"
-              stroke-width="0"
-              viewBox="0 0 1024 1024"
-              className="h-5 w-5 fill-yellow-400"
-              height="1em"
-              width="1em"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M908.1 353.1l-253.9-36.9L540.7 86.1c-3.1-6.3-8.2-11.4-14.5-14.5-15.8-7.8-35-1.3-42.9 14.5L369.8 316.2l-253.9 36.9c-7 1-13.4 4.3-18.3 9.3a32.05 32.05 0 0 0 .6 45.3l183.7 179.1-43.4 252.9a31.95 31.95 0 0 0 46.4 33.7L512 754l227.1 119.4c6.2 3.3 13.4 4.4 20.3 3.2 17.4-3 29.1-19.5 26.1-36.9l-43.4-252.9 183.7-179.1c5-4.9 8.3-11.3 9.3-18.3 2.7-17.5-9.5-33.7-27-36.3z"></path>
-            </svg>
-          </div>
-        </div>
+        <Star5 />
+        <Star4 />
+        <Star3 />
+        <Star2 />
+        <Star1 />
+        <div className="my-4 h-[1px] bg-gray-300"></div>
+        <button
+          type="button"
+          className="flex items-center outline-none transition duration-300 bg-primary text-white flex items-center space-x-2 rounded-md px-4 py-2 hover:bg-opacity-80"
+        >
+          Xóa tất cả
+        </button>
       </div>
     </div>
   );
