+++ Shopping-React/frontEnd/src/components/Star/index.ts
@@ -1,0 +1,5 @@
+// Import named exports
+
+export { Star1, Star2, Star3, Star4, Star5 } from "./Star";
+
+// Export default
