+++ Shopping-React/frontEnd/src/components/TopHome/Tophome.tsx
@@ -1,20 +1,33 @@
-import React from "react";
+import React, { useState } from "react";
+import Button from "./components/Button";
 
 export default function Tophome() {
+  const [activeButton, setActiveButton] = useState(null);
+
+  const handleButtonClick = (buttonLabel) => {
+    setActiveButton(buttonLabel);
+  };
+
   return (
     <div className="bg-gray-300/40 py-4 px-3">
       <div className="flex flex-wrap items-center justify-between gap-2">
         <div className="flex flex-wrap items-center gap-2">
           <div>Sắp xếp theo</div>
-          <button className="h-8 px-4 text-center text-sm capitalize  bg-white text-black hover:bg-slate-100">
-            Phổ biến
-          </button>
-          <button className="h-8 px-4 text-center text-sm capitalize  bg-primary text-white hover:bg-primary/80">
-            Mới nhất
-          </button>
-          <button className="h-8 px-4 text-center text-sm capitalize  bg-white text-black hover:bg-slate-100">
-            Bán chạy
-          </button>
+          <Button
+            label="Phổ biến"
+            isActive={activeButton === "Phổ biến"}
+            onClick={() => handleButtonClick("Phổ biến")}
+          />
+          <Button
+            label="Mới nhất"
+            isActive={activeButton === "Mới nhất"}
+            onClick={() => handleButtonClick("Mới nhất")}
+          />
+          <Button
+            label="Bán chạy"
+            isActive={activeButton === "Bán chạy"}
+            onClick={() => handleButtonClick("Bán chạy")}
+          />
           <select className="h-8 px-4 text-left text-sm capitalize text-black outline-none  bg-white text-black hover:bg-slate-100">
             <option value="" className="bg-white text-black">
               Giá
