+++ Shopping-React/frontEnd/src/index.css
@@ -12,15 +12,7 @@ body {
   font-weight: 400;
 }
 
-@layer utilities {
-  .default-input {
-    @apply block w-full rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-blue-500 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500;
-  }
 
-  .error-input {
-    @apply dark:border-red-400" placeholder=" Bonnie Green"> <p class=" mt-2 block w-full rounded-lg border border-red-500 bg-red-50 p-2.5 text-sm text-red-600 placeholder-red-700 focus:border-red-500 focus:ring-red-500 dark:bg-red-100 dark:text-red-500;
-  }
-}
 
 * {
   margin: 0;
