+++ Shopping-React/frontEnd/vite.config.ts
@@ -4,4 +4,10 @@ import WindiCSS from "vite-plugin-windicss";
 
 export default defineConfig({
   plugins: [react(), WindiCSS()],
+  server: {
+    port: 3000,
+  },
+  preview: {
+    port: 8080,
+  },
 });
